requirements.txt
aioredis==2.0.1
APScheduler==3.9.1
async-timeout==4.0.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
cleo==1.0.0a4
click==8.1.3
crashtest==0.3.1
Deprecated==1.2.13
greenlet==1.1.2
hiredis==2.0.0
idna==3.3
packaging==21.3
pptree==3.1
psycopg2-binary==2.9.3
pyaes==1.6.1
pydantic==1.9.1
pylev==1.4.0
pyparsing==3.0.9
Pyrogram==2.0.26
PySocks==1.7.1
python-decouple==3.6
python-ulid==1.1.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
redis==4.3.3
redis-om==0.0.27
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.37
TgCrypto==1.2.3
types-redis==4.2.6
types-six==1.16.15
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
wrapt==1.14.1
